refactor enterprise dashboard and settings logic to its own app
