+++ corehq/apps/accounting/urls.py
@@ -9,7 +9,6 @@ from corehq.apps.accounting.views import (
     CustomerInvoiceSummaryView,
     EditSoftwarePlanView,
     EditSubscriptionView,
-    EnterpriseBillingStatementsView,
     InvoiceSummaryView,
     ManageAccountingAdminsView,
     ManageBillingAccountView,
@@ -24,12 +23,6 @@ from corehq.apps.accounting.views import (
     SoftwarePlanVersionView,
     WireInvoiceSummaryView,
     accounting_default,
-    edit_enterprise_settings,
-    enterprise_dashboard,
-    enterprise_dashboard_download,
-    enterprise_dashboard_email,
-    enterprise_dashboard_total,
-    enterprise_settings,
     TriggerDowngradeView,
     TriggerAutopaymentsView,
 )
@@ -71,16 +64,3 @@ urlpatterns = [
         name=AccountingAdminInterfaceDispatcher.name()),
 ]
 
-domain_specific = [
-    url(r'^dashboard/$', enterprise_dashboard, name='enterprise_dashboard'),
-    url(r'^dashboard/(?P<slug>[^/]*)/download/(?P<export_hash>[\w\-]+)/$', enterprise_dashboard_download,
-        name='enterprise_dashboard_download'),
-    url(r'^dashboard/(?P<slug>[^/]*)/email/$', enterprise_dashboard_email,
-        name='enterprise_dashboard_email'),
-    url(r'^dashboard/(?P<slug>[^/]*)/total/$', enterprise_dashboard_total,
-        name='enterprise_dashboard_total'),
-    url(r'^settings/$', enterprise_settings, name='enterprise_settings'),
-    url(r'^settings/edit/$', edit_enterprise_settings, name='edit_enterprise_settings'),
-    url(r'^billing_statements/$', EnterpriseBillingStatementsView.as_view(),
-        name=EnterpriseBillingStatementsView.urlname),
-]
