+++ corehq/apps/enterprise/enterprise.py
@@ -10,7 +10,7 @@ from memoized import memoized
 from couchforms.analytics import get_last_form_submission_received
 from dimagi.utils.dates import DateSpan
 
-from corehq.apps.accounting.exceptions import EnterpriseReportError
+from corehq.apps.enterprise.exceptions import EnterpriseReportError
 from corehq.apps.accounting.models import BillingAccount, Subscription
 from corehq.apps.accounting.utils import get_default_domain_url
 from corehq.apps.app_manager.dbaccessors import get_brief_apps_in_domain
