+++ corehq/apps/enterprise/exceptions.py
@@ -1,0 +1,2 @@
+class EnterpriseReportError(Exception):
+    pass
