+++ corehq/apps/enterprise/templates/enterprise/enterprise_dashboard.html
@@ -4,7 +4,7 @@
 
 {% block page_title %}{{ account.name }}{% endblock %}
 
-{% requirejs_main 'accounting/js/enterprise_dashboard' %}
+{% requirejs_main 'enterprise/js/enterprise_dashboard' %}
 
 {% block page_content %}
   {% registerurl "enterprise_dashboard_email" domain "---" %}
