+++ corehq/apps/enterprise/templates/enterprise/enterprise_settings.html
@@ -2,7 +2,7 @@
 {% load crispy_forms_tags %}
 {% load hq_shared_tags %}
 
-{% requirejs_main 'accounting/js/enterprise_settings' %}
+{% requirejs_main 'enterprise/js/enterprise_settings' %}
 
 {% block page_content %}
   {% initial_page_data 'accounts_email' accounts_email %}
